# feat(core): smarter batch action execution