--- a/.changeset/gorgeous-laws-work.md
+++ b/.changeset/gorgeous-laws-work.md
@@ -1,0 +1,5 @@
+---
+'@chugsplash/core': patch
+---
+
+Smarter management of batched action execution