Updates batch action execution with helper functions that figure stuff out automatically to reduce the number of things that need to be hard-coded. Now also batch executes implementation deployments.