@sam-goldman 2022-12-09
Unfortunately, this new approach actually slows down the time that it takes to execute the SetStorage actions by quite a bit if there are a lot of SetStorage actions in the bundle. I timed the execution of the SetStorage actions before and after this PR for Storage.config.ts in the plugins package (which has 59 SetStorage actions) on the Hardhat network:

before: 10.175s
after: 46.702s

However, for Optimism's devnetL1 deployment, which has 21 SetStorage actions, this PR kept the execution time the same:

Optimism (devnetL1 network):
before: 16.284s
after: 16.465s

On the bright side, there was a significant speed-up for the DeployImplementation actions. I only recorded this for Optimism's deployment since Storage.config.ts only contains one contract.

before: 1:28.793 (m:ss.mmm)
after: 44.630s

So, I think it'd be best to keep the naive solution for the SetStorage actions unless there's some way to modify the binary search algorithm so that it performs less `estimateGas` calls (which I think is the issue). We should def use this PR for the DeployImplementation actions since it speeds things up by ~2x.